+++ ONPCalc/MathOperators/AdditionMathOperator.cs
@@ -3,7 +3,7 @@ namespace OnpCalc.MathOperators;
 public class AdditionMathOperator : IMathOperator
 {
     public string Identity => "+";
-    public int Execute(int op1, int op2)
+    public double Execute(double op1, double op2)
     {
         return op1 + op2;
     }
