+++ ONPCalc/NumberSystems/INumberSystem.cs
@@ -4,5 +4,5 @@ namespace OnpCalc.NumberSystems;
 
 public interface INumberSystem
 {
-    bool convert(string data, [NotNullWhen(true)]out int? number);
+    bool convert(string data, [NotNullWhen(true)]out double? number);
 }
